Fixed logic for hashing secret

+++ packages/pronunciations/pronunciation/routes/pronunciations/removePronunciation.js
@@ -1,5 +1,5 @@
 const mysql = require('../../mysql-client');
-const { encryptSecret } = require('../../utils');
+const { compareSecret } = require('../../utils');
 
 async function removePronunciation(args) {
   const id = args.params.pronunciation_id;
@@ -13,21 +13,41 @@ async function removePronunciation(args) {
 
   try {
     const connection = await mysql;
-    const encryptedSecret = await encryptSecret(secret);
+    const selectSql = `
+      SELECT 
+        BIN_TO_UUID(uuid) as uuid,
+        secret
+      FROM pronunciations
+      WHERE uuid = UUID_TO_BIN('?')
+    `
 
-    const sql = `
+    const [selectRow] = await connection.query(selectSql, [id]);
+    const selectedPronunciation = selectRow[0];
+    if (!selectedPronunciation) {
+      return {
+        statusCode: 400,
+        body: `unknown resource with id: ${id} `
+      } 
+    }
+
+    const secretMatches = await compareSecret(secret, selectedPronunciation.secret);
+    if (!secretMatches) {
+      console.log(`unauthed attempt to remove ${id}`)
+      return {
+        statusCode: 401,
+        body: 'Unauthorized'
+      }
+    }
+    
+    const deleteSql = `
       DELETE FROM pronunciations
       WHERE uuid = UUID_TO_BIN(?) AND secret = ?
     `;
 
-    const [results] = await connection.query(sql, [id, encryptedSecret]);
-    if (results.affectedRows === 0) {
-      console.log(`unauthed attempt to remove ${id}`)
-      return {
-        statusCode: 401,
-        body: 'Unauthorized'    
-      }
-    }
+    const [results] = await connection.query(
+      deleteSql, 
+      [id, selectedPronunciation.secret]
+    );
 
     return {
       statusCode: 200,
